wip: cleanup / refactoring
