+++ elixir/web-ng/assets/js/app.js
@@ -28,10 +28,12 @@ import {tableFromIPC} from "apache-arrow"
 import * as d3 from "d3"
 import {sankey as d3Sankey, sankeyLinkHorizontal as d3SankeyLinkHorizontal} from "d3-sankey"
 import mapboxgl from "mapbox-gl"
-import {COORDINATE_SYSTEM, Deck, Layer, OrthographicView, picking, project32} from "@deck.gl/core"
+import {COORDINATE_SYSTEM, Deck, OrthographicView} from "@deck.gl/core"
 import {ArcLayer, LineLayer, ScatterplotLayer, TextLayer} from "@deck.gl/layers"
-import {Geometry, Model} from "@luma.gl/engine"
 import {GodViewWasmEngine} from "./wasm/god_view_exec_runtime"
+import PacketFlowLayer from "./lib/deckgl/PacketFlowLayer"
+import SRQLTimeCookie from "./hooks/SRQLTimeCookie"
+import {nfFormatRateValue} from "./utils/formatters"
 
 import {
   attachTimeTooltip as nfAttachTimeTooltip,
@@ -72,165 +74,6 @@ function getPhoenixTheme() {
     return 'dark';
   }
   return 'light';
-}
-
-function nfFormatBytes(n) {
-  const v = Number(n || 0)
-  if (!Number.isFinite(v)) return "0 B"
-  const abs = Math.abs(v)
-  if (abs >= 1e12) return `${(v / 1e12).toFixed(2)} TB`
-  if (abs >= 1e9) return `${(v / 1e9).toFixed(2)} GB`
-  if (abs >= 1e6) return `${(v / 1e6).toFixed(2)} MB`
-  if (abs >= 1e3) return `${(v / 1e3).toFixed(2)} KB`
-  return `${v.toFixed(0)} B`
-}
-
-function nfFormatBits(n) {
-  const v = Number(n || 0)
-  if (!Number.isFinite(v)) return "0 b"
-  const abs = Math.abs(v)
-  if (abs >= 1e12) return `${(v / 1e12).toFixed(2)} Tb`
-  if (abs >= 1e9) return `${(v / 1e9).toFixed(2)} Gb`
-  if (abs >= 1e6) return `${(v / 1e6).toFixed(2)} Mb`
-  if (abs >= 1e3) return `${(v / 1e3).toFixed(2)} Kb`
-  return `${v.toFixed(0)} b`
-}
-
-function nfFormatCountPerSec(n) {
-  const v = Number(n || 0)
-  if (!Number.isFinite(v)) return "0 /s"
-  const abs = Math.abs(v)
-  if (abs >= 1e9) return `${(v / 1e9).toFixed(2)} G/s`
-  if (abs >= 1e6) return `${(v / 1e6).toFixed(2)} M/s`
-  if (abs >= 1e3) return `${(v / 1e3).toFixed(2)} K/s`
-  return `${v.toFixed(2)} /s`
-}
-
-function nfFormatRateValue(units, n) {
-  const u = String(units || "").trim()
-  if (u === "bps") return `${nfFormatBits(n)}/s`
-  if (u === "Bps") return `${nfFormatBytes(n)}/s`
-  if (u === "pps") return nfFormatCountPerSec(n)
-  return nfFormatBytes(n)
-}
-
-const packetFlowVS = `\
-#define SHADER_NAME sr-packet-flow-layer-vs
-attribute vec2 a_from;
-attribute vec2 a_to;
-attribute float a_seed;
-attribute float a_speed;
-attribute float a_size;
-attribute float a_jitter;
-attribute vec4 a_color;
-
-uniform float u_time;
-
-varying vec4 vColor;
-
-float hash(float n) {
-  return fract(sin(n) * 43758.5453123);
-}
-
-void main(void) {
-  float t = fract((u_time * a_speed) + a_seed);
-  float eased = pow(t, 1.18);
-  vec2 base = mix(a_from, a_to, eased);
-
-  vec2 dir = normalize(max(vec2(0.0001), a_to - a_from));
-  vec2 normal = vec2(-dir.y, dir.x);
-
-  float jitterSeed = hash(a_seed * 91.733);
-  float spread = (jitterSeed - 0.5) * a_jitter;
-  float wobble = sin(u_time * 7.5 + a_seed * 29.0) * (a_jitter * 0.28);
-
-  vec2 pos = base + normal * (spread + wobble);
-
-  vColor = a_color;
-  float tailFade = 1.0 - smoothstep(0.82, 1.0, t);
-  float headBoost = 0.74 + (1.0 - t) * 0.26;
-  vColor.a = clamp(vColor.a * tailFade * headBoost, 0.0, 1.0);
-  gl_Position = project_position_to_clipspace(vec3(pos, 0.0), vec3(0.0), vec3(0.0));
-  gl_PointSize = a_size;
-}
-`
-
-const packetFlowFS = `\
-#define SHADER_NAME sr-packet-flow-layer-fs
-precision highp float;
-varying vec4 vColor;
-
-void main(void) {
-  vec2 p = gl_PointCoord * 2.0 - 1.0;
-  float r = length(p);
-  if (r > 1.0) {
-    discard;
-  }
-  float glow = 1.0 - pow(r, 1.2);
-  float alpha = glow * vColor.a;
-  gl_FragColor = vec4(vColor.rgb, alpha);
-}
-`
-
-class PacketFlowLayer extends Layer {
-  getShaders() {
-    return {vs: packetFlowVS, fs: packetFlowFS, modules: [project32, picking]}
-  }
-
-  initializeState() {
-    const attributeManager = this.getAttributeManager()
-    attributeManager.addInstanced({
-      a_from: {size: 2, accessor: "getFrom"},
-      a_to: {size: 2, accessor: "getTo"},
-      a_seed: {size: 1, accessor: "getSeed"},
-      a_speed: {size: 1, accessor: "getSpeed"},
-      a_size: {size: 1, accessor: "getSize"},
-      a_jitter: {size: 1, accessor: "getJitter"},
-      a_color: {size: 4, type: 5121, normalized: true, accessor: "getColor"},
-    })
-
-    this.setState({
-      model: this._getModel(),
-    })
-  }
-
-  _getModel() {
-    return new Model(this.context.device, {
-      ...this.getShaders(),
-      geometry: new Geometry({
-        topology: "point-list",
-        attributes: {
-          positions: {value: new Float32Array([0, 0, 0]), size: 3},
-        },
-      }),
-      isInstanced: true,
-    })
-  }
-
-  draw({uniforms}) {
-    const model = this.state.model
-    if (!model) return
-    model.setUniforms({
-      ...uniforms,
-      u_time: Number(this.props.time || 0),
-    })
-    model.draw()
-  }
-
-  finalizeState() {
-    this.state.model?.delete?.()
-  }
-}
-
-PacketFlowLayer.defaultProps = {
-  getFrom: (d) => d.from,
-  getTo: (d) => d.to,
-  getSeed: (d) => d.seed,
-  getSpeed: (d) => d.speed,
-  getSize: (d) => d.size,
-  getJitter: (d) => d.jitter,
-  getColor: (d) => d.color,
-  time: 0,
 }
 
 // Custom hooks
@@ -373,113 +216,7 @@ const Hooks = {
     }
   },
 
-  SRQLTimeCookie: {
-    mounted() {
-      this._input = this.el.querySelector('input[name="q"]')
-      if (!this._input) return
-
-      this._debounceTimer = null
-      this._lastSynced = (this.el.dataset.query || "").toString()
-
-      const hasQParam = () => {
-        try {
-          return new URLSearchParams(window.location.search).has("q")
-        } catch (_e) {
-          return false
-        }
-      }
-
-      const cookieGet = (name) => {
-        const needle = `${name}=`
-        const parts = (document.cookie || "").split(";").map((s) => s.trim())
-        for (const part of parts) {
-          if (part.startsWith(needle)) return decodeURIComponent(part.slice(needle.length))
-        }
-        return null
-      }
-
-      const cookieSet = (name, value, days = 365) => {
-        if (!value) return
-        const maxAge = days * 24 * 60 * 60
-        document.cookie = `${name}=${encodeURIComponent(value)}; Max-Age=${maxAge}; Path=/; SameSite=Lax`
-      }
-
-      const extractTimeToken = (q) => {
-        if (!q || typeof q !== "string") return null
-        const m = q.match(/(?:^|\\s)time:(?:\"([^\"]+)\"|(\\S+))/)
-        return m ? (m[1] || m[2] || null) : null
-      }
-
-      const upsertTimeToken = (q, timeToken) => {
-        if (!q || typeof q !== "string") q = ""
-        const trimmed = q.trim()
-        const replacement = ` time:${timeToken}`
-        if (/(?:^|\\s)time:(?:\"[^\"]+\"|\\S+)/.test(trimmed)) {
-          return trimmed.replace(/(?:^|\\s)time:(?:\"[^\"]+\"|\\S+)/, replacement).trim()
-        }
-        return (trimmed + replacement).trim()
-      }
-
-      const persistFromInput = () => {
-        const token = extractTimeToken(this._input.value)
-        if (token) cookieSet("srql_time", token)
-      }
-
-      const maybeRestore = () => {
-        if (hasQParam()) {
-          // Respect deep links; just persist whatever the URL/query contains.
-          persistFromInput()
-          return
-        }
-
-        const token = cookieGet("srql_time")
-        if (!token) return
-
-        const current = (this._input.value || "").toString()
-        const next = upsertTimeToken(current, token)
-        if (next !== current) {
-          this._input.value = next
-          if (typeof this.el.requestSubmit === "function") {
-            this.el.requestSubmit()
-          } else {
-            this.el.dispatchEvent(new Event("submit", { bubbles: true, cancelable: true }))
-          }
-        }
-      }
-
-      this._onInput = () => {
-        clearTimeout(this._debounceTimer)
-        this._debounceTimer = setTimeout(() => persistFromInput(), 150)
-      }
-      this._onSubmit = () => persistFromInput()
-
-      this._input.addEventListener("input", this._onInput)
-      this.el.addEventListener("submit", this._onSubmit)
-
-      maybeRestore()
-    },
-    updated() {
-      // LiveView keeps form inputs "sticky" to preserve user typing, which is usually good.
-      // For SRQL-driven pages, other UI controls can emit SRQL via push_patch, and we want
-      // the topbar query to reflect that new query. Sync it when the input isn't focused.
-      if (!this._input) return
-      if (document.activeElement === this._input) return
-
-      const desired = (this.el.dataset.query || "").toString()
-      if (!desired) return
-
-      const current = (this._input.value || "").toString()
-      if (current !== desired) {
-        this._input.value = desired
-        this._lastSynced = desired
-      }
-    },
-    destroyed() {
-      if (this._input && this._onInput) this._input.removeEventListener("input", this._onInput)
-      if (this._onSubmit) this.el.removeEventListener("submit", this._onSubmit)
-      clearTimeout(this._debounceTimer)
-    }
-  },
+  SRQLTimeCookie,
 
   TimeseriesChart: {
     mounted() {
